added scheduling system

+++ src/main/java/nyu/Meeting.java
@@ -3,13 +3,13 @@ package nyu;
 import utils.JsonSerializable;
 
 public class Meeting implements JsonSerializable {
-  public final String beginDate;     // contains date and time of first event.
-  public final long minutesDuration; // Duration of meeting
-  public final String endDate;       // When the meeting stops repeating
+  public final String beginDate;    // contains date and time of first event.
+  public final int minutesDuration; // Duration of meeting
+  public final String endDate;      // When the meeting stops repeating
 
   public Meeting(String beginDate, String minutesDuration, String endDate) {
     this.beginDate = beginDate;
-    this.minutesDuration = Long.parseLong(minutesDuration);
+    this.minutesDuration = Integer.parseInt(minutesDuration);
     this.endDate = endDate;
   }
